+++ web/packages/teleterm/src/services/config/keyboardShortcutSchema.ts
